Move ngeo layertree.decorate.js into misc.decorate.js
