+++ test/spec/services/decoratelayer.spec.js
@@ -1,15 +1,15 @@
-goog.require('ngeo.layertree.decorate');
+goog.require('ngeo.misc.decorate');
 goog.require('ol.Map');
 goog.require('ol.layer.Tile');
 goog.require('ol.source.OSM');
 
-describe('ngeo.layertree.DecorateLayer', () => {
+describe('ngeo.misc.DecorateLayer', () => {
   it('can change the visibility', () => {
     const layer = new ol.layer.Tile({
       source: new ol.source.OSM(),
       visible: false
     });
-    ngeo.layertree.decorate.layer(layer);
+    ngeo.misc.decorate.layer(layer);
     layer.visible = true;
     expect(layer.getVisible()).toBe(true);
     layer.visible = false;
@@ -21,7 +21,7 @@ describe('ngeo.layertree.DecorateLayer', () => {
       source: new ol.source.OSM(),
       opacity: 0.5
     });
-    ngeo.layertree.decorate.layer(layer);
+    ngeo.misc.decorate.layer(layer);
     layer.opacity = 0.7;
     expect(layer.getOpacity()).toBe(0.7);
   });
